Use size_t instead of long for GPU memory listing
Use size_t instead of long to fix overflow on platforms where long is
only 32-bit (Win64).

+++ libgpujpeg/gpujpeg_common.h
@@ -75,11 +75,11 @@ struct gpujpeg_device_info
     /// Compute capability minor version
     int cc_minor;
     /// Amount of global memory
-    long global_memory;
+    size_t global_memory;
     /// Amount of constant memory
-    long constant_memory;
+    size_t constant_memory;
     /// Amount of shared memory
-    long shared_memory;
+    size_t shared_memory;
     /// Number of registers per block
     int register_count;
     /// Number of multiprocessors
